cfg: abstraction for defining creating a node.
i added yet another field to node, thus necessitating me changing
every damn place i make one with an additional 'nil' argument.
for testing and the like, most of the time when we create a node,
we only care about a couple fields... and we have to update that
code with every field we add.

so let's just have a function that creates nodes and use that.