bfd: Use Go's debug/elf.
This allows us to delete a bunch of our custom code.  Still a bit
far from being able to delete the C entirely.  Dropping the C
would be nice because it means we need a recent-ish version of Go,
since old Cgo's do not support CFLAGS and I'm not willing to write
C89.  Cgo is wicked slow anyway, it would be great to drop it.