cfg: fix depth calculation bug.
in a nested loop situation, the "fail" conditional on a nested
loop was being considered as "under" *both* loop headers, when in
reality it is only technically part of the outer loop.  this uses
a slight modification to the reachability analysis to make sure
only the "relevant" loop headers actually have an effect.