first stab at running our malloc replacement code.
change fspace_fill to only return the end address of the function
it creates, because it now inserts code at a *given* address
instead of inventing an address for the user.  then use it in some
test code that "jump"s to the function it creates and hacks
thereby changes malloc's operation.