Implement access detection based on DSM ideas.
This is the 'meat' of the access detection scheme.  We change our
allocation routine ("tjfmalloc") so that it also mprotects the
memroy to be PROT_READ only.  Then we wait for a segfault: when it
happens, we check the list of allocations we know about, and it's
a segfault that we caused due to our mprotect'ing, we unprotect
it, clear the signal so the application never receives it, and
continue on.

In the future, we'll have to call "vis(...)" on the chunk of
memory that was accessed when this happens.  We'll also need to
recreate the read-only protection, so that we'll be informed of
subequent accesses as well.