working(?) access detection.
it works for 20 or 30 iterations, at least.  Then it gets some
crazy error reading the inferior's set of libraries.

This is probably due to a memory corruption bug, but we don't know
where yet.  The quickest solution is probably to just start
caching symbols, since they don't change and the memory corruption
is likely in the code that loads those, anyway.