normalize using a uniform.
need to find the uniform's location in Pre(), for some reason.  GL
thinks the uniform name or the program (ambiguous which) is
invalid if I wait until Render().  Strange.  Anyway, I guess the
uniform location should be constant, so this should be faster...

This also doesn't quite seem to work; the texture coordinate
lookup is wrapping badly, or something else awful is happening.
can't quite tell at the moment.