bfd: change 'needed' to be library-level, not symbol-level.
There are too many cases where we need the symbol and we can't
know until someone tests with a given program, otherwise.

The issue is that we need to get pretty exact functions from
find_function in the case that we have detected an access.  This
is because we need to filter accesses: not all accesses give us
information, and, worse, we can't build a control flow graph for
e.g. 'calloc', because DynInst reads the data from the executable
file, not the process' memory.

Thus we need to check to see if the access is in a
cannot-and-should-not-build-CFG function and explicitly skip it.
If we're missing symbols, we won't (accurately) identify what
function we are in, and will mess up that identification.

Right now, our ID is based purely on a hardcoded list of functions
that I noticed applications hit but are not actually relevant for
access.  This is a terrible way to do it.