try a bit harder to find the return address.
before resorting to the heavy-handed and often-problematic
approach of searching through the instruction stream for the next
'ret'.  The search tends to work poorly because our symbol list is
not guaranteed to be perfect.  This is an issue because we try to
search from the start of the function we are in, not the current
instruction we are at (in case the return is above is in the
function), and if we think we are in the wrong function then,
well, we can find the wrong 'return' (or even not find one at
all).