more hacks for 'which functions are interesting'.
this approach is flawed.  we need it right now, however, perhaps
because we're catching all accesses of heap memory.  We should
really be filtering things so that, for example, we don't look at
memory that 'fopen' allocates internally.  Thresholding whether or
not we watch memory based on how large the allocation was is
probably also very useful.  Data assigned into a 'char*' is
probably similarly useless: we don't care what strings the user
code handles.

Once we start accurately restricting ourselves to memory regions
that are actually of relevance, it should lower the accesses we
see.  With that restriction, we probably won't need to rely so
heavily on this kind of filtering (though I don't see how we'd
ever be able to remove it outright).