InfEvents: keep a set of callbacks per breakpoint.
I don't think we actually need an actual set, because I believe
every time we want to 'overwrite' the callback pointer, we are
actually overwriting it with a pointer to the *same* callback.

Unfortunately, Go does not allow one to compare function pointers,
so there is no way to force that belief via runtime checks.
Instead, we just keep a set; I guess it could theoretically be
useful to have breakpoints for different kinds of things at the
same instruction address.