switch to go-gl-legacy, update python for API change.
apparently go-gl worsened itself by getting rid of all the nice
typing stuff that it used to have.  at least they moved it to
go-gl-legacy.

the python package's "InitModule" now requires one to supply the
array of methods.  I guess this might be the exported methods...
hopefully this works.  I think we only expose some variables, not
methods, and so hopefully this doesn't really matter.