"fix" the 2D case.
I can't set the damn uniform on my computer, unless I do it in
Pre().  So I just hacked it to a proper max for linpack, which is
of course total garbage in general.

this seems to be an issue with just my laptop, will need to
revisit on a new machine...